fix(schemas:powerstream): Update schema restrictions in setCommand.ts
The parameter restrictions were altered in several schemas in the setCommand.ts file. For customLoadPowerSettingsSchema, the requirement for the permanentWatts parameter to be an integer was removed. For upperChargingLevelSchema and indicatorLightBrightnessSchema, the parameter names were changed from 'lowerLimit' to 'upperLimit' and 'brightness' respectively. For powerStreamDeleteTaskSchema, the maximum limit for the taskIndex was removed.