feat(schemas:glacier): Add setCommands schemas for Glacier package
This commit introduces a new file that defines several schemas for the setCommands of the Glacier package. This allows for the setting of various properties of the Glacier device such as temperature, ECO mode, buzzer enabling status, screen timeout, and more. Additionally, updated the index files and the README of the schema package to reflect this addition.